Adds pre-git code
All the junk I did at work today. Includes a basic G4P UI, a JSONHelper
class to supplement JSONObject, a Palette class which stores an
ArrayList of Colors and handles IO to/from Strings, and a Channel class
which holds a base Palette and an ArrayList of corresponding Palettes
that can be swapped in.